pyproject.toml
[project]
name = "converter-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "fastapi>=0.124.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]
